pyproject.toml
[tool.poetry]
name = "markets-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.7.2,<3.10"
click = "^7.1.2"
fastapi = "^0.64.0"
eth-brownie = "^1.14.6"
pymongo = "^3.11.4"
python-jose = {extras = ["cryptography"], version = "^3.2.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.dev-dependencies]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
